remove arrow_legacy.py file
